Display patient procedures on report
